fix(server): serverless deploy fixed
