+++ packages/server/package.json
@@ -3,10 +3,6 @@
 	"version": "1.0.0",
 	"description": "Darkmoon Serverless API",
 	"main": "handler.js",
-	"eslintConfig": {
-		"extends": "@serverless/eslint-config/node",
-		"root": true
-	},
 	"scripts": {
 		"test": "echo \"Error: no test specified\" && exit 1"
 	},
@@ -15,17 +11,15 @@
 		"source-map-support": "^0.5.10"
 	},
 	"devDependencies": {
-		"serverless-plugin-monorepo": "^0.9.0",
 		"@types/aws-lambda": "^8.10.17",
 		"@types/jsonwebtoken": "^8.5.0",
 		"@types/serverless": "^1.72.5",
-		"@types/webpack-node-externals": "^2.5.0",
 		"aws-sdk": "^2.786.0",
-		"fork-ts-checker-webpack-plugin": "^3.0.1",
-		"git-list-updated": "^1.2.1",
 		"serverless": "^2.10.0",
-		"webpack": "^4.29.0",
-		"webpack-node-externals": "^1.7.2"
+		"serverless-offline": "^6.8.0",
+		"serverless-plugin-monorepo": "^0.9.0",
+		"serverless-plugin-typescript": "^1.1.9",
+		"@darkmoon/typings": "*"
 	},
 	"license": "MIT"
 }
