+++ packages/server/serverless.ts
@@ -17,7 +17,7 @@ const serverlessConfiguration: Serverless = {
 		}
 	},
 	// Add the serverless-webpack plugin
-	plugins: ['serverless-webpack'],
+	plugins: ['serverless-plugin-typescript', 'serverless-offline'],
 	provider: {
 		name: 'aws',
 		runtime: 'nodejs12.x',
