+++ packages/server/src/handlers/info.ts
@@ -1,5 +1,4 @@
 import { APIGatewayProxyHandler } from 'aws-lambda'
-import 'source-map-support/register'
 import infoHandler from '../controllers/info'
 
 export const getInfo: APIGatewayProxyHandler = async (event, context) => {
