+++ packages/server/tsconfig.json
@@ -1,24 +1,16 @@
 {
-	"compilerOptions": {
-		"experimentalDecorators": true,
-		"lib": ["ESNext"],
-		"removeComments": true,
-		"moduleResolution": "node",
-		"noUnusedLocals": true,
-		"noUnusedParameters": true,
+  "compilerOptions": {
+    "allowSyntheticDefaultImports": true,
+    "esModuleInterop": true,
+    "lib": ["esnext"],
+    "module": "commonjs",
+    "moduleResolution": "node",
+    "noUnusedLocals": true,
+    "noUnusedParameters": true,
+    "outDir": "lib",
 		"sourceMap": true,
-		"target": "es2017",
-		"outDir": "lib",
-		"rootDir": "/server"
-	},
-	"include": ["src/**/*.ts"],
-	"exclude": [
-		"node_modules/**",
-		"node_modules/**/*",
-		".serverless/**/*",
-		".webpack/**/*",
-		"_warmup/**/*",
-		".vscode/**/*"
-	]
+		"experimentalDecorators": true,
+    "target": "esnext"
+  },
+  "exclude": ["node_modules"]
 }
-
