+++ packages/typings/Product.ts
@@ -1,10 +1,12 @@
 import * as yup from 'yup'
+import { ProductImage } from './ProductImage'
 
 export type Product = {
 	id: string
 	name: string
 	description: string
 	price: number
+	images: ProductImage[]
 }
 
 export const productCreateSchema = yup.object().shape({
@@ -17,6 +19,6 @@ export const productGetSchema = yup.object().shape({
 	productId: yup.string().required()
 })
 
-export type ProductCreateResponse = {
+export type ProductCreatePayload = {
 	objectLocation: string
 }
