+++ packages/typings/ProductImage.ts
