+++ packages/typings/package.json
@@ -1,0 +1,6 @@
+{
+  "name": "@darkmoon/typings",
+  "version": "1.0.0",
+  "main": "index.ts",
+  "license": "MIT"
+}
