--- packages/web/src/types/models/Product.ts
@@ -1,8 +1,0 @@
-import { ProductImage } from './ProductImage'
-
-export interface Product {
-	id: number
-	images: ProductImage[]
-	name: string
-	price: string
-}
